splitted websocket code
