+++ lib/runner/cli/cli.js
@@ -12,7 +12,7 @@ const analyticsCollector = require('../../utils/analytics.js');
 const {Logger, singleSourceFile} = Utils;
 const {RealIosDeviceIdError, iosRealDeviceUDID, isRealIos, isMobile, killSimulator} = require('../../utils/mobile.js');
 const {Browser} = require('selenium-webdriver');
-const SelectorPlaygroundServer = require('../../testsuite/playground-listener.js');
+const SelectorPlaygroundServer = require('../../testsuite/selector-playground');
 
 class CliRunner {
   static get CONFIG_FILE_JS() {
